+++ splmeter/processing/time.py
@@ -138,8 +138,8 @@ class TimeWeight(BaseModule):
         integration_time_index_size = int(input_fs*self.output_resolution)
         if integration_time_index_size <= 0:
             integration_time_index_size = 1
-        start_index = self.start_time*input_fs
-        buffer = max(0,(integration_window_index_size - start_index))
+        start_index = int(self.start_time*input_fs)
+        buffer = int(max(0,(integration_window_index_size - start_index)))
         start_index += buffer
 
         sig_arr=np.concatenate([np.array([0]*buffer),sig_arr],axis=0)
